Overhauled Backend
Overhauled Backend:
    -  All DB operations happen through server
    -  Server Infrastructure Redesigned
    -  Tidied up Dashboard.js and Timer.js
    -  Tidied up directory
