--- .gitignore
@@ -1,3 +1,0 @@
-node_modules/
-bun.lockb
-package-lock.json
