--- README.md
@@ -1,5 +1,0 @@
-# Software-Development-AS91906
-
-This program is designed for use by both University and High School students to
-help manage study, and ensure they are meeting the study goals they 
-set out for themselves.
