--- package.json
@@ -1,5 +1,0 @@
-{
-    "dependencies": {
-        "express": "^4.19.2"
-    }
-}
